continue scraping, detect last chapter

+++ scraper.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python
 
+import errno
 import os
+import os.path
 import sys
 
 import requests
@@ -14,12 +16,16 @@ bounds = {
 	),
 	'book2': (
 		'https://practicalguidetoevil.wordpress.com/2015/11/04/prologue-2/',
-		'https://practicalguidetoevil.wordpress.com/2016/10/12/heroic-interlude-prise-au-fer/'
+		None
 	),
 }
 start, end = bounds[book]
 dirname = 'pgte_%s_raw/' % book
-os.mkdir(dirname)
+try:
+	os.mkdir(dirname)
+except OSError as e:
+	if e.errno != errno.EEXIST:
+		raise
 
 rs = requests.Session()
 url = start
@@ -27,11 +33,19 @@ i = 0
 while url != end:
 	name = url.rsplit('/', 2)[1]
 	filename = '%02d-%s' % (i, name)
-	print 'getting', filename
-	content = rs.get(url).content
-	with open(dirname + filename, 'w') as f:
-		f.write(content)
+	if os.path.exists(dirname + filename):
+		print 'already have', filename
+		with open(dirname + filename, 'r') as f:
+			content = f.read()
+	else:
+		print 'getting', filename
+		content = rs.get(url).content
+		with open(dirname + filename, 'w') as f:
+			f.write(content)
 	soup = BeautifulSoup(content, 'lxml')
 	next_el = soup.find('link', rel='next')
-	url = next_el['href']
+	if next_el is None:
+		url = None
+	else:
+		url = next_el['href']
 	i += 1
